microservice/backend/requirements_celery.txt
pdf2image==1.14.0
opencv_python==4.4.0.44
numpy==1.18.5
Pillow==8.0.1
celery==5.0.2
redis==3.5.3